+++ core/dialer.go
@@ -9,15 +9,22 @@ import (
 	"io"
 	"log"
 	"net"
+	"net/http"
 	"strconv"
 )
 
-func Dialer(logger *log.Logger, pipingServerUrl string, tcpPort uint16, path string) error {
+func Dialer(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, httpHeaders [][]string, networkType NetworkType, port uint16, path string) error {
 	logger.Printf("answer-side")
 	errCh := make(chan error)
 
-	// Create a new RTCPeerConnection
-	peerConnection, err := NewDetachablePeerConnection(createConfig())
+	var peerConnection *webrtc.PeerConnection
+	var err error
+	if networkType == NetworkTypeTcp {
+		peerConnection, err = NewDetachablePeerConnection(createConfig())
+	} else {
+		// NOTE: UDP does not need to detach
+		peerConnection, err = webrtc.NewPeerConnection(createConfig())
+	}
 	if err != nil {
 		return err
 	}
@@ -43,7 +50,24 @@ func Dialer(logger *log.Logger, pipingServerUrl string, tcpPort uint16, path str
 			errCh <- nil
 		}
 	})
+	switch networkType {
+	case NetworkTypeTcp:
+		tcpDialer(logger, peerConnection, port)
+	case NetworkTypeUdp:
+		udpDialer(logger, peerConnection, port)
+	}
 
+	go func() {
+		answer := piping_webrtc_signaling.NewAnswer(logger, httpClient, pipingServerUrl, httpHeaders, peerConnection, answerSideId(path), offerSideId(path))
+		if err := answer.Start(); err != nil {
+			errCh <- err
+		}
+	}()
+
+	return <-errCh
+}
+
+func tcpDialer(logger *log.Logger, peerConnection *webrtc.PeerConnection, port uint16) {
 	// Register data channel creation handling
 	peerConnection.OnDataChannel(func(d *webrtc.DataChannel) {
 		logger.Printf("OnDataChannel")
@@ -55,7 +79,7 @@ func Dialer(logger *log.Logger, pipingServerUrl string, tcpPort uint16, path str
 				logger.Printf("failed to detach: %+v", err)
 				return
 			}
-			conn, err := net.Dial("tcp", ":"+strconv.Itoa(int(tcpPort)))
+			conn, err := net.Dial("tcp", ":"+strconv.Itoa(int(port)))
 			if err != nil {
 				logger.Printf("failed to dial", err)
 				raw.Close()
@@ -65,13 +89,40 @@ func Dialer(logger *log.Logger, pipingServerUrl string, tcpPort uint16, path str
 			go io.Copy(conn, raw)
 		})
 	})
+}
 
-	go func() {
-		answer := piping_webrtc_signaling.NewAnswer(logger, pipingServerUrl, peerConnection, answerSideId(path), offerSideId(path))
-		if err := answer.Start(); err != nil {
-			errCh <- err
+func udpDialer(logger *log.Logger, peerConnection *webrtc.PeerConnection, port uint16) {
+	// Register data channel creation handling
+	peerConnection.OnDataChannel(func(d *webrtc.DataChannel) {
+		logger.Printf("OnDataChannel")
+		// TODO: hard code: 127.0.0.1
+		conn, err := net.Dial("udp", net.JoinHostPort("127.0.0.1", strconv.Itoa(int(port))))
+		if err != nil {
+			log.Printf("failed to dial: %v", err)
+			d.Close()
+			return
 		}
-	}()
+		// Register channel opening handling
+		d.OnOpen(func() {
+			logger.Printf("data channel OnOpen")
+			var buf [65536]byte
+			for {
+				n, err := conn.Read(buf[:])
+				if err != nil {
+					logger.Printf("failed to read: %+v", err)
+					return
+				}
+				if err := d.Send(buf[:n]); err != nil {
+					logger.Printf("failed to send: %+v", err)
+					return
+				}
+			}
+		})
 
-	return <-errCh
+		d.OnMessage(func(msg webrtc.DataChannelMessage) {
+			if _, err := conn.Write(msg.Data); err != nil {
+				logger.Printf("failed to write: %+v", err)
+			}
+		})
+	})
 }