+++ piping-webrtc-signaling/common.go
@@ -14,14 +14,22 @@ type InitialJson struct {
 	Version uint64 `json:"version"`
 }
 
-func sendSdp(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, localId string, remoteId string, description *webrtc.SessionDescription) error {
+func sendSdp(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, httpHeaders [][]string, localId string, remoteId string, description *webrtc.SessionDescription) error {
 	payload, err := json.Marshal(description)
 	if err != nil {
 		return err
 	}
 	url := fmt.Sprintf("%s/%s-%s/sdp", pipingServerUrl, localId, remoteId)
 	logger.Printf("sending sdp to %s...", url)
-	resp, err := httpClient.Post(url, "application/json; charset=utf-8", bytes.NewReader(payload))
+	req, err := http.NewRequest("POST", url, bytes.NewReader(payload))
+	if err != nil {
+		return err
+	}
+	req.Header.Set("Content-Type", "application/json; charset=utf-8")
+	for _, kv := range httpHeaders {
+		req.Header.Add(kv[0], kv[1])
+	}
+	resp, err := httpClient.Do(req)
 	if err != nil {
 		return err
 	}
@@ -34,10 +42,17 @@ func sendSdp(logger *log.Logger, httpClient *http.Client, pipingServerUrl string
 	return nil
 }
 
-func receiveSdp(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, localId string, remoteId string) (*webrtc.SessionDescription, error) {
+func receiveSdp(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, httpHeaders [][]string, localId string, remoteId string) (*webrtc.SessionDescription, error) {
 	url := fmt.Sprintf("%s/%s-%s/sdp", pipingServerUrl, remoteId, localId)
 	logger.Printf("receiving sdp from %s ...", url)
-	r, err := httpClient.Get(url)
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, err
+	}
+	for _, kv := range httpHeaders {
+		req.Header.Add(kv[0], kv[1])
+	}
+	r, err := httpClient.Do(req)
 	if err != nil {
 		return nil, err
 	}
@@ -48,13 +63,21 @@ func receiveSdp(logger *log.Logger, httpClient *http.Client, pipingServerUrl str
 	return &sdp, nil
 }
 
-func sendCandidate(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, localId string, remoteId string, c *webrtc.ICECandidate) error {
+func sendCandidate(logger *log.Logger, httpClient *http.Client, pipingServerUrl string, httpHeaders [][]string, localId string, remoteId string, c *webrtc.ICECandidate) error {
 	candidateBytes, err := json.Marshal(c.ToJSON())
 	if err != nil {
 		return err
 	}
 	logger.Printf("sending candidate...")
-	resp, err := httpClient.Post(fmt.Sprintf("%s/%s-%s/candidate", pipingServerUrl, localId, remoteId), "application/json; charset=utf-8", bytes.NewReader(candidateBytes))
+	req, err := http.NewRequest("POST", fmt.Sprintf("%s/%s-%s/candidate", pipingServerUrl, localId, remoteId), bytes.NewReader(candidateBytes))
+	if err != nil {
+		return err
+	}
+	req.Header.Set("Content-Type", "application/json; charset=utf-8")
+	for _, kv := range httpHeaders {
+		req.Header.Add(kv[0], kv[1])
+	}
+	resp, err := httpClient.Do(req)
 	if err != nil {
 		return err
 	}
@@ -67,8 +90,15 @@ func sendCandidate(logger *log.Logger, httpClient *http.Client, pipingServerUrl 
 	return nil
 }
 
-func receiveCandidate(httpClient *http.Client, pipingServerUrl string, localId string, remoteId string) (*webrtc.ICECandidateInit, error) {
-	res, err := httpClient.Get(fmt.Sprintf("%s/%s-%s/candidate", pipingServerUrl, remoteId, localId))
+func receiveCandidate(httpClient *http.Client, pipingServerUrl string, httpHeaders [][]string, localId string, remoteId string) (*webrtc.ICECandidateInit, error) {
+	req, err := http.NewRequest("GET", fmt.Sprintf("%s/%s-%s/candidate", pipingServerUrl, remoteId, localId), nil)
+	if err != nil {
+		return nil, err
+	}
+	for _, kv := range httpHeaders {
+		req.Header.Add(kv[0], kv[1])
+	}
+	res, err := httpClient.Do(req)
 	if err != nil {
 		return nil, err
 	}