+++ CHANGELOG.md
@@ -5,8 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-07-04
+### Added
+* Support UDP port forwarding
+* Add --insecure option
+* Add --dns-server option
+* Add --header option
+
 ## 0.1.0 - 2022-07-02
 ### Added
 * Initial release
 
-[Unreleased]: https://github.com/nwtgck/go-webrtc-piping-tunnel/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/nwtgck/go-webrtc-piping-tunnel/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/nwtgck/go-webrtc-piping-tunnel/compare/v0.1.0...v0.2.0